# samples/server/petstore/jaxrs-spec-interface/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Swagger Petstore",
    "description" : "This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "license" : {
      "name" : "Apache-2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
...
  "servers" : [ {
    "url" : "http://petstore.swagger.io:80/v2"
  } ],
...
  "paths" : {
...
    "/dog" : {
      "put" : {
        "tags" : [ "dog" ],
        "summary" : "Update an existing dog",
        "operationId" : "updateDog",
        "requestBody" : {
          "description" : "Dog object that needs to be added.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Dog"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/Dog"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied",
            "content" : { }
          },
          "404" : {
            "description" : "Animal not found",
            "content" : { }
          },
          "405" : {
            "description" : "Validation exception",
            "content" : { }
          }
        }
      },
      "post" : {
        "tags" : [ "dog" ],
        "summary" : "Add a new dog to the store",
        "operationId" : "addDog",
        "requestBody" : {
          "description" : "Dog object that needs to be added to the store",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Dog"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/Dog"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "405" : {
            "description" : "Invalid input",
            "content" : { }
          }
        }
      }
    },
    "/dog/{dogId}" : {
      "get" : {
        "tags" : [ "dog" ],
        "summary" : "Find dog by ID",
        "description" : "Returns a single dog",
        "operationId" : "getDogById",
        "parameters" : [ {
          "name" : "dogId",
          "in" : "path",
          "description" : "ID of dog to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Dog"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Dog"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid ID supplied",
            "content" : { }
          },
          "404" : {
            "description" : "Pet not found",
            "content" : { }
          }
        }
      },
      "post" : {
        "tags" : [ "dog" ],
        "summary" : "Updates a dog",
        "operationId" : "updateDogWithForm",
        "parameters" : [ {
          "name" : "dogId",
          "in" : "path",
          "description" : "ID of dog that needs to be updated",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_7"
              }
            }
          }
        },
        "responses" : {
          "405" : {
            "description" : "Invalid input",
            "content" : { }
          }
        }
      },
      "delete" : {
        "tags" : [ "dog" ],
        "summary" : "Deletes a dog",
        "operationId" : "deleteDog",
        "parameters" : [ {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dogId",
          "in" : "path",
          "description" : "Dog id to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid dog value",
            "content" : { }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Dog" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Animal"
        }, {
          "type" : "object",
          "properties" : {
            "breed" : {
              "type" : "string"
            }
          }
        } ]
      },
...
      "Animal" : {
        "required" : [ "className" ],
        "type" : "object",
        "properties" : {
          "className" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string",
            "default" : "red"
          }
        },
        "discriminator" : {
          "propertyName" : "className"
        }
      },
...
      "body_7" : {
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "Updated name of the dog"
          },
          "status" : {
            "type" : "string",
            "description" : "Updated status of the dog"
          }
        }
      }
    },
    "securitySchemes" : {
      "petstore_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "http://petstore.swagger.io/api/oauth/dialog",
            "scopes" : {
              "write:pets" : "modify pets in your account",
              "read:pets" : "read your pets"
            }
          }
        }
      },
      "http_basic_test" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      },
      "api_key_query" : {
        "type" : "apiKey",
        "name" : "api_key_query",
        "in" : "query"
      }
    }
  }
}